Refactor package structure. Extracted table constants.
