+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/home/model/repository/local/lastfm/sqldb/LastFMTableConstants.kt
@@ -1,0 +1,14 @@
+package ayds.lisboa.songinfo.moredetails.home.model.repository.local.lastfm.sqldb
+
+const val ARTISTS_TABLE = "artists"
+const val ID_COLUMN = "id"
+const val ARTIST_COLUMN = "artist"
+const val INFO_COLUMN = "info"
+const val SOURCE_COLUMN = "source"
+
+const val createArtistTableQuery: String =
+    "create table $ARTISTS_TABLE (" +
+            "$ID_COLUMN integer PRIMARY KEY, " +
+            "$ARTIST_COLUMN string, " +
+            "$INFO_COLUMN string, " +
+            "$SOURCE_COLUMN integer)"
